add cmake for kinetis k32l2 family
